Fix clippy errors

+++ src/utils/mod.rs
@@ -264,7 +264,7 @@ impl<T: Digest + Debug + Write + Clone> EifBuilder<T> {
     }
 
     fn kernel_size(&self) -> u64 {
-        self.kernel.metadata().unwrap().len() as u64
+        self.kernel.metadata().unwrap().len()
     }
 
     fn cmdline_offset(&self) -> u64 {
@@ -289,7 +289,7 @@ impl<T: Digest + Debug + Write + Clone> EifBuilder<T> {
     }
 
     fn ramdisk_size(&self, ramdisk: &File) -> u64 {
-        ramdisk.metadata().unwrap().len() as u64
+        ramdisk.metadata().unwrap().len()
     }
 
     fn signature_offset(&self) -> u64 {
